API-546 separate loader to another module. Not sure it will work
